LayoutTests/fast/table/auto-layout-close-to-max-width-expected.txt

PASS document.getElementById("onlyTable").offsetWidth is 999999
PASS successfullyParsed is true

TEST COMPLETE
